Reduce code duplication tenfold
