+++ src/sodiumCompatibility/java/net/coderbot/iris/compat/sodium/impl/vertex_format/entity_xhfp/writer/EntityVertexBufferWriterUnsafe.java
@@ -4,12 +4,12 @@ import me.jellysquid.mods.sodium.client.model.vertex.buffer.VertexBufferView;
 import me.jellysquid.mods.sodium.client.model.vertex.buffer.VertexBufferWriterUnsafe;
 import net.coderbot.iris.compat.sodium.impl.vertex_format.IrisModelVertexFormats;
 import net.coderbot.iris.compat.sodium.impl.vertex_format.entity_xhfp.EntityVertexSink;
+import net.coderbot.iris.compat.sodium.impl.vertex_format.entity_xhfp.QuadViewEntity;
 import net.coderbot.iris.vertices.NormalHelper;
-import net.coderbot.iris.vertices.QuadView;
 import org.lwjgl.system.MemoryUtil;
 
 public class EntityVertexBufferWriterUnsafe extends VertexBufferWriterUnsafe implements EntityVertexSink {
-	private final QuadView quad = new QuadView();
+	private final QuadViewEntity quad = new QuadViewEntity();
 	private final int STRIDE;
 
 	public EntityVertexBufferWriterUnsafe(VertexBufferView backingBuffer) {
@@ -42,7 +42,11 @@ public class EntityVertexBufferWriterUnsafe extends VertexBufferWriterUnsafe imp
 
 		quad.setup(writePointer, STRIDE);
 
-		computeTangents(i, length, normalX, normalY, normalZ);
+		int tangent = quad.computeTangent(normalX, normalY, normalZ);
+
+		for (long vertex = 0; vertex < length; vertex++) {
+			MemoryUtil.memPutInt(i - 4 - STRIDE * vertex, tangent);
+		}
 
 		float midU = 0;
 		float midV = 0;
@@ -58,103 +62,6 @@ public class EntityVertexBufferWriterUnsafe extends VertexBufferWriterUnsafe imp
 		for (long vertex = 0; vertex < length; vertex++) {
 			MemoryUtil.memPutFloat(i - 12 - STRIDE * vertex, midU);
 			MemoryUtil.memPutFloat(i - 8 - STRIDE * vertex, midV);
-		}
-	}
-
-	private void computeTangents(long i, int length, float normalX, float normalY, float normalZ) {
-		// Capture all of the relevant vertex positions
-		float x0 = quad.x(0);
-		float y0 = quad.y(0);
-		float z0 = quad.z(0);
-
-		float x1 = quad.x(1);
-		float y1 = quad.y(1);
-		float z1 = quad.z(1);
-
-		float x2 = quad.x(2);
-		float y2 = quad.y(2);
-		float z2 = quad.z(2);
-
-		float edge1x = x1 - x0;
-		float edge1y = y1 - y0;
-		float edge1z = z1 - z0;
-
-		float edge2x = x2 - x0;
-		float edge2y = y2 - y0;
-		float edge2z = z2 - z0;
-
-		float u0 = quad.u(0);
-		float v0 = quad.v(0);
-
-		float u1 = quad.u(1);
-		float v1 = quad.v(1);
-
-		float u2 = quad.u(2);
-		float v2 = quad.v(2);
-
-		float deltaU1 = u1 - u0;
-		float deltaV1 = v1 - v0;
-		float deltaU2 = u2 - u0;
-		float deltaV2 = v2 - v0;
-
-		float fdenom = deltaU1 * deltaV2 - deltaU2 * deltaV1;
-		float f;
-
-		if (fdenom == 0.0) {
-			f = 1.0f;
-		} else {
-			f = 1.0f / fdenom;
-		}
-
-		float tangentx = f * (deltaV2 * edge1x - deltaV1 * edge2x);
-		float tangenty = f * (deltaV2 * edge1y - deltaV1 * edge2y);
-		float tangentz = f * (deltaV2 * edge1z - deltaV1 * edge2z);
-		float tcoeff = rsqrt(tangentx * tangentx + tangenty * tangenty + tangentz * tangentz);
-		tangentx *= tcoeff;
-		tangenty *= tcoeff;
-		tangentz *= tcoeff;
-
-		float bitangentx = f * (-deltaU2 * edge1x + deltaU1 * edge2x);
-		float bitangenty = f * (-deltaU2 * edge1y + deltaU1 * edge2y);
-		float bitangentz = f * (-deltaU2 * edge1z + deltaU1 * edge2z);
-		float bitcoeff = rsqrt(bitangentx * bitangentx + bitangenty * bitangenty + bitangentz * bitangentz);
-		bitangentx *= bitcoeff;
-		bitangenty *= bitcoeff;
-		bitangentz *= bitcoeff;
-
-		// predicted bitangent = tangent × normal
-		// Compute the determinant of the following matrix to get the cross product
-		//  i  j  k
-		// tx ty tz
-		// nx ny nz
-
-		float pbitangentx =   tangenty * normalZ - tangentz * normalY;
-		float pbitangenty = -(tangentx * normalZ - tangentz * normalX);
-		float pbitangentz =   tangentx * normalX - tangenty * normalY;
-
-		float dot = (bitangentx * pbitangentx) + (bitangenty * pbitangenty) + (bitangentz * pbitangentz);
-		float tangentW;
-
-		if (dot < 0) {
-			tangentW = -1.0F;
-		} else {
-			tangentW = 1.0F;
-		}
-
-		int tangent = NormalHelper.packNormal(tangentx, tangenty, tangentz, tangentW);
-
-		for (long vertex = 0; vertex < length; vertex++) {
-			MemoryUtil.memPutInt(i - 4 - STRIDE * vertex, tangent);
-		}
-	}
-
-	private static float rsqrt(float value) {
-		if (value == 0.0f) {
-			// You heard it here first, folks: 1 divided by 0 equals 1
-			// In actuality, this is a workaround for normalizing a zero length vector (leaving it as zero length)
-			return 1.0f;
-		} else {
-			return (float) (1.0 / Math.sqrt(value));
 		}
 	}
 }
